+++ API/CancellationTokenSource.js
@@ -7,6 +7,13 @@ class CancellationTokenSource {
 	constructor() {
 		/**@type {string} */
 		this.Token = uuidv4();
+		this._cancel = false
+	}
+	Cancel(){
+		this._cancel = true
+	}
+	IsCancellationRequested(){
+		return this._cancel
 	}
 }
 module.exports = {